PowerPont reporting of a PCA

PowerPont reporting of a PCA

Yes_PCA(
  res,
  yes_study_name = "PCA + HCPC",
  path = getwd(),
  file.name = "PCA_results.pptx",
  x1 = 1,
  x2 = 2,
  proba = 0.05,
  size_tab = 10
)

Arguments

res
a PCA object

yes_study_name
title displayed on the first slide

path
PowerPoint file to be created

file.name
name of the PowerPoint file

x1
component to plot on the x-axis

x2
component to plot on the y-axis

proba
the significance threshold considered to characterized the category (by default 0.05)

size_tab
maximum number of rows of a table per slide

Value

Returns a .pptx file

Examples

if (FALSE) {
library(FactoMineR)
data(decathlon)
res.pca <- FactoMineR::PCA(decathlon, quanti.sup = 11:12, quali.sup=13)
# Create the PowerPoint in the current working directory
Yes_PCA(res.pca)
}